move tests to proper groups
