+++ src/test/java/com/wikia/webdriver/testcases/wikiamobiletests/MobileEditModeTests.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.testcases.mobile;
+package com.wikia.webdriver.testcases.wikiamobiletests;
 
 import com.wikia.webdriver.common.contentpatterns.MobilePageContent;
 import com.wikia.webdriver.common.contentpatterns.PageContent;
